re design wip
