+++ Client/src/app/Core/ToolBars/Inside/inside-toolbar.component.ts
@@ -3,6 +3,7 @@ import { ToolBarService } from '../../Services/tool-bar.service';
 import { Router } from '@angular/router';
 import { MatBottomSheet } from '@angular/material/bottom-sheet';
 import { SettingsComponent } from 'src/app/Main/Components/Settings/settings.component';
+import { MovieDto } from 'src/app/Data/Interfaces/movie.dto';
 
 @Component({
   selector: 'inside-toolbar',
@@ -23,6 +24,30 @@ export class InsideToolbarComponent {
      */
     getIsLoading(): boolean {
         return this._toolBarService.isLoading; 
+    }
+
+    /**
+     * Gets an array of numbers representing the years from 1900 to present.
+     */
+    getYears(): number[] {
+        var years = [];
+        var currentYear = new Date().getFullYear();
+        for (var i = currentYear; i > 1900; i--){
+            years.push(i);
+        }
+        return years;
+    }
+
+    
+    /**
+     * Toggles the movies list and the No Movies Found response in view.
+     * @param movies 
+     */
+    moviesExist(movies: MovieDto[], search: string) : boolean {
+        if (!search) {
+            return false;
+        }
+        return movies && movies.length > 0; 
     }
 
     /**
