+++ Client/src/app/Main/Components/Movie/movie.component.scss
@@ -4,13 +4,13 @@
         flex-direction: row;
         align-items: center;
         justify-content: center;
-        height: 500px;
+        height: 700px;
         width: 900px;
         padding: 0px;
         background-color: rgb(59, 1, 1);
         .movie-poster {
-            height: 500px;
-            width: 40%;
+            height: 700px;
+            width: 50%;
         }
 
         .info-container {
