+++ Client/src/app/Main/Components/Recommendations/recommendations.component.html
@@ -1,4 +1,5 @@
 <div id="movie-finder" *ngIf="!isError(error)">
+    <h1 class="title">Discover</h1>
     <div class="recs-container">
         <div class="inline-movie-arrow">
             <button mat-button (click)="changeMovie(-1)">
