+++ Client/src/app/Main/Components/SearchHistory/movie-carousel.component.scss
@@ -7,7 +7,7 @@
         display: flex;
         align-items: center;
         flex-direction: column;
-        padding: 30px;
+        padding: 5px;
  
         .history-container {
             display: flex;
@@ -17,8 +17,8 @@
                 color: white;
             }
             img {
-                height: 300px;
-                width: 200px;
+                height: 200px;
+                width: 125px;
                 margin: 5px;
                 cursor: pointer;
                 border: 3px solid white;
